refactor: add features folder and move some files to it
