+++ src/app/auth/registration-email-resending/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { LinkExpiredForm } from '@/src/app/auth/registration-email-resending/link-expired-form/LinkExpiredForm'
+import { LinkExpiredForm } from '@/src/features/link-expired-form/LinkExpiredForm'
 
 export default function LinkExpiredPage() {
   return <LinkExpiredForm />
