+++ src/app/auth/registration/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { RegistrationForm } from '@/src/app/auth/registration/registration-form/RegistrationForm'
+import { RegistrationForm } from '@/src/features/registration-form/RegistrationForm'
 
 export default function RegistrationPage() {
   return <RegistrationForm />
