+++ src/features/registration-confirmation-form/email-confirmed.module.scss
